requirements.txt
# NEXUS OVERLORD v2.0 - Python Dependencies
# Erstellt: 2025-11-23

# Web Framework
flask>=3.0.0
werkzeug>=3.0.0

# Datenbank
# sqlite3 ist built-in in Python

# KI APIs (OpenRouter für Gemini + Sonnet)
requests>=2.31.0
httpx>=0.24.0

# PDF Export
reportlab>=4.0.0

# Dokument-Extraktion (PDF/DOCX Upload)
pdfplumber>=0.10.0
python-docx>=1.0.0

# Markdown Rendering
markdown2>=2.4.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
